+++ previous learning stuff/learnApp.js
